ref(dev): Share contracts with mock API
Define pure oRPC contracts for routes with mock handlers. Use the same contracts for production and mock implementations.

Compose the mock router through one implementer so TypeScript checks route nesting without loading production handlers.
