+++ apps/server/package.json
@@ -25,6 +25,7 @@
     "@logtape/logtape": "catalog:",
     "@logtape/redaction": "catalog:",
     "@opentelemetry/api": "catalog:",
+    "@orpc/contract": "catalog:",
     "@orpc/openapi": "catalog:",
     "@orpc/server": "catalog:",
     "@orpc/zod": "catalog:",
