+++ apps/server/src/orpc/index.ts
@@ -1,17 +1,7 @@
 import { os } from "@orpc/server";
 import sentryMiddleware from "@peated/orpc/server/middleware";
 import type { User } from "../db/schema";
-
-interface ErrorContract {
-  // the message should always be a punctuated sentence
-  message: string;
-  // internal error code if needed
-  code?: string;
-  // mapped to HTTP status code
-  statusCode: number;
-  // mapped to HTTP status text
-  error: string;
-}
+import { errorDefinitions } from "./contracts/base";
 
 export const base = os
   .$context<{
@@ -22,53 +12,6 @@ export const base = os
   /**
    * All errors must follow the ErrorContract interface.
    */
-  .errors({
-    // [string: code]: ErrorContract
-    UNAUTHORIZED: {
-      message: "Unauthorized.",
-      statusCode: 401,
-      error: "Unauthorized",
-    },
-    NOT_FOUND: {
-      message: "The resource was not found.",
-      statusCode: 404,
-      error: "Not Found",
-    },
-    BAD_REQUEST: {
-      message: "Bad Request.",
-      statusCode: 400,
-      error: "Bad Request",
-    },
-    FORBIDDEN: {
-      message: "Forbidden.",
-      statusCode: 403,
-      error: "Forbidden",
-    },
-    INTERNAL_SERVER_ERROR: {
-      message: "Internal server error.",
-      statusCode: 500,
-      error: "Internal Server Error",
-    },
-    METHOD_NOT_ALLOWED: {
-      message: "Method not allowed.",
-      statusCode: 405,
-      error: "Method Not Allowed",
-    },
-    CONFLICT: {
-      message: "Conflict.",
-      statusCode: 409,
-      error: "Conflict",
-    },
-    PAYLOAD_TOO_LARGE: {
-      message: "Payload too large.",
-      statusCode: 413,
-      error: "Payload Too Large",
-    },
-    UNPROCESSABLE_ENTITY: {
-      message: "Unprocessable entity.",
-      statusCode: 422,
-      error: "Unprocessable Entity",
-    },
-  });
+  .errors(errorDefinitions);
 
 export const procedure = base.use(sentryMiddleware());
