+++ apps/server/src/orpc/mock/context.ts
@@ -1,6 +1,6 @@
-import type { Outputs } from "@peated/server/orpc/router";
+import type { MockOutputs } from "./contract";
 
 // Use this context whenever production behavior depends on the signed-in user.
 export type MockContext = {
-  user: Outputs["auth"]["login"]["user"] | null;
+  user: MockOutputs["auth"]["login"]["user"] | null;
 };
