+++ apps/server/src/orpc/mock/implementer.ts
@@ -1,0 +1,5 @@
+import { implement } from "@orpc/server";
+import type { MockContext } from "./context";
+import { mockContract } from "./contract";
+
+export const mockOS = implement(mockContract).$context<MockContext>();
