+++ apps/server/src/orpc/mock/routes/bottles/details.ts
@@ -1,17 +1,15 @@
-import { implement } from "@orpc/server";
-import type { MockContext } from "@peated/server/orpc/mock/context";
 import {
   mockBottleDetails,
   mockBottleDetailsFor,
 } from "@peated/server/orpc/mock/fixtures";
-import details from "@peated/server/orpc/routes/bottles/details";
+import { mockOS } from "@peated/server/orpc/mock/implementer";
 
-export default implement(details)
-  .$context<MockContext>()
-  .handler(async ({ input, context, errors }) => {
+export default mockOS.bottles.details.handler(
+  async ({ input, context, errors }) => {
     if (input.bottle !== mockBottleDetails.id) {
       throw errors.NOT_FOUND({ message: "Mock bottle not found." });
     }
 
     return mockBottleDetailsFor(context.user);
-  });
+  },
+);
