+++ apps/server/src/orpc/mock/routes/entities/list.ts
@@ -1,12 +1,11 @@
-import { implement } from "@orpc/server";
 import {
   includesQuery,
   mockEntity,
   noMorePages,
 } from "@peated/server/orpc/mock/fixtures";
-import list from "@peated/server/orpc/routes/entities/list";
+import { mockOS } from "@peated/server/orpc/mock/implementer";
 
-export default implement(list).handler(async ({ input }) => ({
+export default mockOS.entities.list.handler(async ({ input }) => ({
   results: includesQuery(input.query, mockEntity.name, mockEntity.shortName)
     ? [mockEntity]
     : [],
