+++ apps/server/src/orpc/routes/entities/list.ts
@@ -1,5 +1,6 @@
+import { implement } from "@orpc/server";
 import { parseDetailsFromName } from "@peated/bottle-classifier/smws";
-import { ENTITY_TYPE_LIST } from "@peated/server/constants";
+import sentryMiddleware from "@peated/orpc/server/middleware";
 import { db } from "@peated/server/db";
 import {
   bottles,
@@ -13,70 +14,16 @@ import {
   plainTextSearchQuery,
   prefixTextSearchQuery,
 } from "@peated/server/lib/search";
-import { procedure } from "@peated/server/orpc";
-import { EntitySchema, listResponse } from "@peated/server/schemas";
+import type { Context } from "@peated/server/orpc/context";
+import entityListContract from "@peated/server/orpc/contracts/entities/list";
 import { serialize } from "@peated/server/serializers";
 import { EntitySerializer } from "@peated/server/serializers/entity";
 import type { SQL } from "drizzle-orm";
 import { and, asc, desc, eq, ilike, or, sql } from "drizzle-orm";
-import { z } from "zod";
 
-const DEFAULT_SORT = "rank";
-
-const SORT_OPTIONS = [
-  "rank",
-  "name",
-  "created",
-  "tastings",
-  "bottles",
-  "-name",
-  "-created",
-  "-tastings",
-  "-bottles",
-] as const;
-
-const InputSchema = z
-  .object({
-    query: z
-      .string()
-      .default("")
-      .describe("Plain-text search; operator syntax is not supported."),
-    name: z.string().nullish(),
-    country: z.coerce.string().nullish().describe("Country slug or id"),
-    region: z.coerce.string().nullish().describe("Region slug or id"),
-    type: z.enum(ENTITY_TYPE_LIST).nullish(),
-    bottler: z.number().nullish(),
-    searchContext: z
-      .object({
-        type: z.enum(ENTITY_TYPE_LIST).nullish(),
-        brand: z.number().nullish(),
-        bottleName: z.string().nullish(),
-      })
-      .nullish(),
-    sort: z.enum(SORT_OPTIONS).default(DEFAULT_SORT),
-    cursor: z.coerce.number().gte(1).default(1),
-    limit: z.coerce.number().lte(500).default(100),
-  })
-  .default({
-    query: "",
-    sort: DEFAULT_SORT,
-    cursor: 1,
-    limit: 100,
-  });
-
-const OutputSchema = listResponse(EntitySchema);
-
-export default procedure
-  .route({
-    method: "GET",
-    path: "/entities",
-    summary: "List entities",
-    description:
-      "Search and filter entities (brands, distilleries, bottlers) with advanced filtering by location, type, and search context",
-    operationId: "listEntities",
-  })
-  .input(InputSchema)
-  .output(OutputSchema)
+export default implement(entityListContract)
+  .$context<Context>()
+  .use(sentryMiddleware())
   .handler(async function ({
     input: { query, cursor, limit, ...input },
     context,
