+++ apps/server/src/orpc/routes/users/details.ts
@@ -1,3 +1,5 @@
+import { implement } from "@orpc/server";
+import sentryMiddleware from "@peated/orpc/server/middleware";
 import { db } from "@peated/server/db";
 import {
   bottles,
@@ -8,12 +10,11 @@ import {
 } from "@peated/server/db/schema";
 import { getUserFromId } from "@peated/server/lib/api";
 import { RESERVED_COLLECTIONS } from "@peated/server/lib/db";
-import { procedure } from "@peated/server/orpc";
-import { detailsResponse, UserSchema } from "@peated/server/schemas";
+import type { Context } from "@peated/server/orpc/context";
+import userDetailsContract from "@peated/server/orpc/contracts/users/details";
 import { serialize } from "@peated/server/serializers";
 import { UserSerializer } from "@peated/server/serializers/user";
 import { and, eq, gt, sql } from "drizzle-orm";
-import { z } from "zod";
 import {
   readJoinedUserBottle,
   scanUserTastingBottles,
@@ -94,38 +95,9 @@ async function aggregateCollectionStats(userId: number) {
   return { collected: bottleIds.size, library };
 }
 
-export default procedure
-  .route({
-    method: "GET",
-    path: "/users/{user}",
-    summary: "Get user details",
-    description:
-      "Retrieve user profile information including statistics for tastings, bottles, and contributions",
-    operationId: "getUser",
-  })
-  .input(
-    z.object({
-      user: z.union([z.coerce.number(), z.literal("me"), z.string()]),
-    }),
-  )
-  // TODO(response-envelope): wrap in { data } by updating detailsResponse() at cutover
-  .output(
-    detailsResponse(
-      UserSchema.extend({
-        stats: z.object({
-          tastings: z.number(),
-          bottles: z.number(),
-          collected: z.number(),
-          library: z.object({
-            total: z.number(),
-            open: z.number(),
-            sealed: z.number(),
-          }),
-          contributions: z.number(),
-        }),
-      }),
-    ),
-  )
+export default implement(userDetailsContract)
+  .$context<Context>()
+  .use(sentryMiddleware())
   .handler(async function ({ input, context, errors }) {
     const user = await getUserFromId(db, input.user, context.user);
