+++ pnpm-workspace.yaml
@@ -16,6 +16,7 @@ allowBuilds:
 
 catalog:
   '@orpc/client': 1.15.0
+  '@orpc/contract': 1.15.0
   '@orpc/openapi': 1.15.0
   '@orpc/server': 1.15.0
   '@orpc/shared': 1.15.0
